requirements.txt
Pillow>=10.0
PyQt6>=6.5
openwakeword>=0.6.0
PyAudio>=0.2.14
openai-whisper>=20231117
pyautogui>=0.9.54
pyttsx3>=2.99
